% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaks.R
\name{endemic_countries_by_disease}
\alias{endemic_countries_by_disease}
\title{Endemic Countries of a Disease}
\usage{
endemic_countries_by_disease(disease)
}
\arguments{
\item{disease}{GIDEON disease code}
}
\value{
Returns list of all endemic countries of a disease.
}
\description{
Endemic Countries of a Disease
}
\seealso{
\url{https://api-doc.gideononline.com/#20bc3345-8c15-4cc7-a6a3-24c8ac0ccf49}

\code{\link{lookup_gideon_id}}
}
