% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_diseases.R
\name{filter_diseases}
\alias{filter_diseases}
\title{Filter diseases from GIDEON database}
\usage{
filter_diseases(
  agent = NULL,
  vector = NULL,
  vehicle = NULL,
  reservoir = NULL,
  country = NULL
)
}
\description{
Filter diseases from GIDEON database
}
\seealso{
\code{\link{lookup_gideon_id}}
}
