% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaks.R
\name{latest_outbreaks_by_country}
\alias{latest_outbreaks_by_country}
\title{Latest Disease Outbreaks by Country}
\usage{
latest_outbreaks_by_country(country)
}
\arguments{
\item{country}{GIDEON country code}
}
\value{
Returns information of latest outbreak of every disease for a
requested country.
}
\description{
Latest Disease Outbreaks by Country
}
\examples{
albania_country_code <- lookup_gideon_id("countries", "Albania")
latest_outbreaks_by_country(albania_country_code)
}
\seealso{
\url{https://api-doc.gideononline.com/#a331d79d-c265-48a3-a45a-dbca49c8d38d}

\code{\link{lookup_gideon_id}}
}
