% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_gideon_id.R
\name{lookup_agent_id}
\alias{lookup_agent_id}
\title{GIDEON Agent Code Lookup}
\usage{
lookup_agent_id(agent)
}
\arguments{
\item{agent}{Name of agent}
}
\description{
Shortcut for \code{lookup_gideon_id("agents", agent)}
}
\seealso{
\code{\link{lookup_gideon_id}}
}
