% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_gideon_id.R
\name{lookup_bacteria_id}
\alias{lookup_bacteria_id}
\title{GIDEON Bacteria Code Lookup}
\usage{
lookup_bacteria_id(bacteria)
}
\arguments{
\item{bacteria}{Name of bacteria}
}
\description{
Shortcut for \code{lookup_gideon_id("bacteria", bacteria)}
}
\seealso{
\code{\link{lookup_gideon_id}}
}
