% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_gideon_id.R
\name{lookup_country_id}
\alias{lookup_country_id}
\title{GIDEON Country Code Lookup}
\usage{
lookup_country_id(country)
}
\arguments{
\item{country}{Name of country}
}
\description{
Shortcut for \code{lookup_gideon_id("countries", country)}
}
\seealso{
\code{\link{lookup_gideon_id}}
}
