% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_gideon_id.R
\name{lookup_disease_id}
\alias{lookup_disease_id}
\title{GIDEON Disease Code Lookup}
\usage{
lookup_disease_id(disease)
}
\arguments{
\item{disease}{Name of disease}
}
\description{
Shortcut for \code{lookup_gideon_id("diseases", disease)}
}
\seealso{
\code{\link{lookup_gideon_id}}
}
