% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_gideon_id.R
\name{lookup_drug_id}
\alias{lookup_drug_id}
\title{GIDEON Drug Code Lookup}
\usage{
lookup_drug_id(drug)
}
\arguments{
\item{drug}{Name of drug}
}
\description{
Shortcut for \code{lookup_gideon_id("drugs", drug)}
}
\seealso{
\code{\link{lookup_gideon_id}}
}
