% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_gideon_id.R
\name{lookup_gideon_id}
\alias{lookup_gideon_id}
\title{Finds the GIDEON specific ID for a disease, vaccine, etc.}
\usage{
lookup_gideon_id(category, item)
}
\arguments{
\item{category}{One of the following: diseases, drugs, vaccines, agents,
vectors, vehicles, reservoirs, bacteria, mycobacteria, yeasts,
countries, regions}

\item{item}{The name of the item}
}
\description{
Finds the GIDEON specific ID for a disease, vaccine, etc.
}
