% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_gideon_id.R
\name{lookup_mycobacteria_id}
\alias{lookup_mycobacteria_id}
\title{GIDEON Mycobacteria Code Lookup}
\usage{
lookup_mycobacteria_id(mycobacteria)
}
\arguments{
\item{mycobacteria}{Name of mycobacteria}
}
\description{
Shortcut for \code{lookup_gideon_id("mycobacteria", mycobacteria)}
}
\seealso{
\code{\link{lookup_gideon_id}}
}
