% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_gideon_id.R
\name{lookup_region_id}
\alias{lookup_region_id}
\title{GIDEON Region Code Lookup}
\usage{
lookup_region_id(region)
}
\arguments{
\item{region}{Name of region}
}
\description{
Shortcut for \code{lookup_gideon_id("regions", region)}
}
\seealso{
\code{\link{lookup_gideon_id}}
}
