% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_gideon_id.R
\name{lookup_reservoir_id}
\alias{lookup_reservoir_id}
\title{GIDEON Reservoir Code Lookup}
\usage{
lookup_reservoir_id(reservoir)
}
\arguments{
\item{reservoir}{Name of reservoir}
}
\description{
Shortcut for \code{lookup_gideon_id("reservoirs", reservoir)}
}
\seealso{
\code{\link{lookup_gideon_id}}
}
