% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_gideon_id.R
\name{lookup_vaccine_id}
\alias{lookup_vaccine_id}
\title{GIDEON Vaccine Code Lookup}
\usage{
lookup_vaccine_id(vaccine)
}
\arguments{
\item{vaccine}{Name of vaccine}
}
\description{
Shortcut for \code{lookup_gideon_id("vaccines", vaccine)}
}
\seealso{
\code{\link{lookup_gideon_id}}
}
