% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_gideon_id.R
\name{lookup_vector_id}
\alias{lookup_vector_id}
\title{GIDEON Vector Code Lookup}
\usage{
lookup_vector_id(vector)
}
\arguments{
\item{vector}{Name of vector}
}
\description{
Shortcut for \code{lookup_gideon_id("vectors", vector)}
}
\seealso{
\code{\link{lookup_gideon_id}}
}
