% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_gideon_id.R
\name{lookup_vehicle_id}
\alias{lookup_vehicle_id}
\title{GIDEON Vehicle Code Lookup}
\usage{
lookup_vehicle_id(vehicle)
}
\arguments{
\item{vehicle}{Name of vehicle}
}
\description{
Shortcut for \code{lookup_gideon_id("vehicles", vehicle)}
}
\seealso{
\code{\link{lookup_gideon_id}}
}
