% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_gideon_id.R
\name{lookup_yeast_id}
\alias{lookup_yeast_id}
\title{GIDEON Yeast Code Lookup}
\usage{
lookup_yeast_id(yeast)
}
\arguments{
\item{yeast}{Name of yeast}
}
\description{
Shortcut for \code{lookup_gideon_id("yeasts", yeast)}
}
\seealso{
\code{\link{lookup_gideon_id}}
}
