% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaks.R
\name{outbreaks_by_country_year}
\alias{outbreaks_by_country_year}
\title{Disease Outbreaks by Country and Year}
\usage{
outbreaks_by_country_year(country, year)
}
\arguments{
\item{country}{GIDEON country code}

\item{year}{Numeric. 4 digit year.}
}
\value{
Returns complete list of all outbreaks that were reported in a
requested country and year.
}
\description{
Disease Outbreaks by Country and Year
}
\examples{
us_country_code <- lookup_gideon_id("countries", "United States")
outbreaks_by_country_year(us_country_code, 1980)
}
\seealso{
\url{https://api-doc.gideononline.com/#16616aab-57f9-477d-b611-892b7f19cd43}

\code{\link{lookup_gideon_id}}
}
