% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaks.R
\name{outbreaks_by_year}
\alias{outbreaks_by_year}
\title{Disease Outbreaks by Year}
\usage{
outbreaks_by_year(year)
}
\arguments{
\item{year}{Numeric. 4 digit year.}
}
\value{
Returns a complete list of all outbreaks that occurred in a requested
year for every country.
}
\description{
Disease Outbreaks by Year
}
\examples{
outbreaks_by_year(1980)
}
\seealso{
\url{https://api-doc.gideononline.com/#038e2b12-0b61-4432-a642-7192a64f8a5c}
}
