% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{to_geojson}
\alias{to_geojson}
\title{Dataframe export as GeoJSON}
\usage{
to_geojson(
  dataframe,
  filename,
  overwrite = FALSE,
  pretty = FALSE,
  lat = "latitude",
  lon = "longitude"
)
}
\arguments{
\item{dataframe}{An input dataframe to convert}

\item{filename}{The exported filename, relative to the current working
directory. If the file already exists and overwrite is not set to TRUE,
an error will be thrown}

\item{overwrite}{Indicates if an existing file in the current directory
should be overwritten by the new file. Defaults to FALSE.}

\item{pretty}{Indicates if file is formatted with whitespace as newlines,
indentations and spaces to make the JSON data more readable.
Defaults to FALSE.}

\item{lat}{The name of the column with latitude data. Defaults to "latitude".}

\item{lon}{The name of the column with longitude data. Defaults to "longitude".}
}
\description{
Exports a dataframe to a GeoJSON file where latitude/longitude data is
available.
}
