% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaks.R
\name{endemic_diseases_by_country}
\alias{endemic_diseases_by_country}
\title{Diseases Endemic to a Country}
\usage{
endemic_diseases_by_country(country)
}
\arguments{
\item{country}{GIDEON country code}
}
\value{
Returns the list of all diseases that are endemic to the country.
}
\description{
Diseases Endemic to a Country
}
\seealso{
\url{https://api-doc.gideononline.com/#feab6f00-e926-49df-9a34-114cfafdf685}

\code{\link{lookup_gideon_id}}
}
