% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_diseases.R
\name{filter_diseases}
\alias{filter_diseases}
\title{Filter diseases from GIDEON database}
\usage{
filter_diseases(
  agent = NULL,
  vector = NULL,
  vehicle = NULL,
  reservoir = NULL,
  country = NULL
)
}
\arguments{
\item{agent}{GIDEON agent code - Classification (e.g., virus, parasite) and
taxonomic designation.}

\item{vector}{GIDEON vector code - An arthropod or other living carrier which
transports an infectious agent from an infected organism or reservoir to a
susceptible individual or immediate surroundings.}

\item{vehicle}{The mode of transmission for an infectious agent. This
generally implies a passive and inanimate (i.e., non-vector) mode.}

\item{reservoir}{Any animal, arthropod, plant, soil or substance in which an
infectious agent normally lives and multiplies, on which it depends
primarily for survival, and where it reproduces itself in such a manner that
it can be transmitted to a susceptible host.}

\item{country}{In general, this list of country names corresponds to accepted
geographical and political designations. When Country filter is used, it
will only retrieve diseases that are endemic to the country.}
}
\description{
Filter diseases from GIDEON database
}
\seealso{
\url{https://api-doc.gideononline.com/#4272d285-ba04-435d-b7ad-deabe971330e}

\code{\link{lookup_gideon_id}}
}
