% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_gideon_api.R
\name{query_gideon_api}
\alias{query_gideon_api}
\title{Queries the GIDEON Informatics API}
\usage{
query_gideon_api(path, query = NULL, to_dataframe = TRUE)
}
\arguments{
\item{path}{The API path according to the API documentation}

\item{query}{Additional URL query parameters to send as a named list}

\item{to_dataframe}{Automatically convert the response to a dataframe.}
}
\description{
Queries the GIDEON Informatics API
}
