% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_gideon_api.R
\name{query_gideon_api}
\alias{query_gideon_api}
\title{GIDEON API Query}
\usage{
query_gideon_api(path, query = NULL, to_dataframe = TRUE)
}
\arguments{
\item{path}{The API path according to the API documentation}

\item{query}{Additional URL query parameters to send as a named list}

\item{to_dataframe}{Automatically convert the response to a dataframe.}
}
\value{
A dataframe representing the response or a string of the JSON response if
\code{to_dataframe} is set to \code{FALSE}
}
\description{
Wrapper to the GIDEON API
}
\examples{
# Most calls only need the path
query_gideon_api("/diseases/10050/incubation-periods")

# Some calls can include a named list to be used as query parameters
query_gideon_api("/diseases/filter",
                 list(agent=1014,
                      vector="T",
                      vehicle="MI",
                      reservoir="A",
                      country="G100"))

}
\seealso{
\url{https://api-doc.gideononline.com/}
}
