% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{to_geojson}
\alias{to_geojson}
\title{Dataframe export as GeoJSON}
\usage{
to_geojson(
  dataframe,
  filename,
  overwrite_layer = FALSE,
  lat = "latitude",
  lon = "longitude"
)
}
\arguments{
\item{dataframe}{An input dataframe to convert}

\item{filename}{The exported filename, relative to the current working
directory}

\item{overwrite_layer}{Indicates if an existing file in the current directory
should be overwritten by the new file}

\item{lat}{The name of the column with latitude data}

\item{lon}{The name of the column with longitude data}
}
\description{
Exports a dataframe to GeoJSON where latitude/longitude data is available
}
